+++ libhesai/src/auto_tick_count.cc
@@ -113,7 +113,7 @@ int TickCount::ShowTimeSlice(std::string sLogFile, bool bSaveSysTime) {
 
     if (pFileSel != NULL) fprintf(pFileSel, "%s:\n", iter->first.c_str());
 
-    for (int i = 0; i < iter->second.size(); i++) {
+    for (size_t i = 0; i < iter->second.size(); i++) {
       std::cout << "  " << static_cast<double>(iter->second.at(i)) / kMicroToSec
                << "s\n";
